Removed files from gitignore
Command to do this is

    git rm --cached

The command ls-files shows only tracked files.  The -i option will filter files that are tracked
and that match the patterns included in the file specified with the --exclude-from option.

This ls-files command produces a list of files and we git rm --cached these files.